+++ projects/valtimo/form-link/src/lib/components/choose-process-link-type/choose-process-link-type.component.ts
@@ -15,7 +15,7 @@
  */
 
 import {Component} from '@angular/core';
-import {ProcessLinkState2Service} from '../../services';
+import {ProcessLinkStateService} from '../../services';
 
 @Component({
   selector: 'valtimo-choose-process-link-type',
@@ -26,7 +26,7 @@ export class ChooseProcessLinkTypeComponent {
   public readonly availableProcessLinkTypes$ =
     this.processLinkState2Service.availableProcessLinkTypes$;
 
-  constructor(private readonly processLinkState2Service: ProcessLinkState2Service) {}
+  constructor(private readonly processLinkState2Service: ProcessLinkStateService) {}
 
   selectProcessLinkType(processLinkTypeId: string): void {
     this.processLinkState2Service.selectProcessLinkType(processLinkTypeId);
