+++ projects/valtimo/form-link/src/lib/components/process-link-modal/process-link-modal.component.html
@@ -27,7 +27,7 @@
     showBackButton: showBackButton$ | async,
     showNextButton: showNextButton$ | async,
     enableNextButton: enableNextButton$ | async,
-    hasOneOption: hasOneOption$ | async,
+    hasOneProcessLinkType: hasOneProcessLinkType$ | async,
     saving: saving$ | async
   } as obs"
 >
@@ -51,7 +51,7 @@
 
 <ng-template #progressIndicator let-obs="obs">
   <ibm-progress-indicator
-    *ngIf="obs.steps && !obs.hasOneOption"
+    *ngIf="obs.steps && !obs.hasOneProcessLinkType"
     [steps]="obs.steps"
     [current]="obs.currentStepIndex"
     spacing="equal"
@@ -62,17 +62,22 @@
   <div class="ibm-loading-container" *ngIf="obs.saving">
     <ibm-loading></ibm-loading>
   </div>
-  <ng-container *ngIf="obs.currentStepId === 'chooseProcessLinkType'">
-    <valtimo-choose-process-link-type></valtimo-choose-process-link-type>
-  </ng-container>
-  <ng-container *ngIf="obs.currentStepId === 'selectForm'">
-    <valtimo-select-form></valtimo-select-form>
-  </ng-container>
-  <ng-container *ngIf="obs.currentStepId === 'selectFormFlow'">
-    <valtimo-select-form-flow></valtimo-select-form-flow>
-  </ng-container>
-  <ng-container *ngIf="obs.currentStepId === 'choosePluginConfiguration'">
-    <valtimo-select-plugin-configuration></valtimo-select-plugin-configuration>
+  <ng-container *ngIf="obs.currentStepId as stepId">
+    <ng-container *ngIf="stepId === 'chooseProcessLinkType'">
+      <valtimo-choose-process-link-type></valtimo-choose-process-link-type>
+    </ng-container>
+    <ng-container *ngIf="stepId === 'selectForm'">
+      <valtimo-select-form></valtimo-select-form>
+    </ng-container>
+    <ng-container *ngIf="stepId === 'selectFormFlow'">
+      <valtimo-select-form-flow></valtimo-select-form-flow>
+    </ng-container>
+    <ng-container *ngIf="stepId === 'choosePluginConfiguration'">
+      <valtimo-select-plugin-configuration></valtimo-select-plugin-configuration>
+    </ng-container>
+    <ng-container *ngIf="stepId === 'configurePluginAction'">
+      <valtimo-select-plugin-action></valtimo-select-plugin-action>
+    </ng-container>
   </ng-container>
 </ng-template>
