+++ projects/valtimo/form-link/src/lib/components/process-link/process-link.component.html
@@ -40,7 +40,7 @@
           <valtimo-select-plugin-configuration></valtimo-select-plugin-configuration>
         </v-stepper-step>
         <v-stepper-step titleTranslationKey="pluginSteps.step2">
-          <valtimo-select-plugin-function></valtimo-select-plugin-function>
+          <valtimo-select-plugin-action></valtimo-select-plugin-action>
         </v-stepper-step>
         <v-stepper-step titleTranslationKey="pluginSteps.step3">
           <valtimo-plugin-function-configuration
