+++ projects/valtimo/form-link/src/lib/services/index.ts
@@ -15,9 +15,9 @@
  */
 
 export * from './form-flow.service';
-export * from './process-link-state.service';
+export * from './plugin-state.service';
 export * from './form-link.service';
 export * from './process-link.service';
-export * from './process-link-state-2.service';
+export * from './process-link-state.service';
 export * from './process-link-step.service';
 export * from './process-link-button.service';
