+++ projects/valtimo/form-link/src/public-api.ts
@@ -23,7 +23,7 @@ export * from './lib/services/form-link.service';
 export * from './lib/components/form-link/form-link.component';
 export * from './lib/components/process-link/process-link.component';
 export * from './lib/components/select-plugin-configuration/select-plugin-configuration.component';
-export * from './lib/components/select-plugin-function/select-plugin-function.component';
+export * from './lib/components/select-plugin-action/select-plugin-action.component';
 export * from './lib/components/plugin-function-configuration/plugin-function-configuration.component';
 export * from './lib/form-link.module';
 export * from './lib/models';
